review di detail photographer sudah bisa

+++ app/src/main/java/com/tugasakhir/photographerbooking/model/pojo/Review.kt
@@ -5,6 +5,6 @@ import kotlinx.android.parcel.Parcelize
 
 @Parcelize
 data class Review (
-    val score: Float,
+    val score: Int,
     val review: String,
 ): Parcelable
